requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
requests==2.31.0
reportlab==4.0.4
fpdf==1.7.2
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
Werkzeug==2.3.6
python-owasp-zap-v2.4==0.0.22
openai==1.3.0